use one translator file for 2.0 and 2.1
